blog-service/requirements.txt
Django==5.0
djangorestframework==3.15
psycopg2-binary
django-redis
django-filter
python-slugify
gunicorn
drf-spectacular
drf-spectacular[sidecar]